Replaced implicit cast from long to int with explicit cast

+++ nifi-extension-bundles/nifi-standard-bundle/nifi-standard-processors/src/main/java/org/apache/nifi/processors/standard/SplitPcap.java
@@ -192,7 +192,7 @@ public class SplitPcap extends AbstractProcessor {
 
             } else {
                 newPackets.add(packet);
-                currentPacketCollectionSize += (packet.inclLen() + packetHeaderLength);
+                currentPacketCollectionSize += ((int) packet.inclLen() + packetHeaderLength);
                 unprocessedPackets.remove(0);
             }
         }
